Shortened GUID prefixes to avoid the following warnings when validating with Orca:
   ICE03 Warnings: String overflow (greater than length permitted in column)

"The Identifier data type is a text string. Identifiers may contain the ASCII characters A-Z (a-z), digits, underscores (_), or periods (.). However, every identifier must begin with either a letter or an underscore."

http://msdn.microsoft.com/en-us/library/aa369212(v=vs.85).aspx

The identifier allows a maximum of 72 characters, which seems undocumented. This limitation is met when creating merge modules out of fragments produced with Paraffin.